Chunk reclamation feature for slot mem pool. Change limit.
The original implementation of slot mem pool has no ability
to reuse chunk for slots size X if it was split for slots size Y.
It can cause the situation when we have no abbility to allocate memory
it all chunks was splited for wirkset with different size.

The proposed change track usage of chunk and allows to reuse it
for any slot size if use count for chunk become 0.

This commit also changes limit so allocate object up to 32MB are allowed.

Two improvement should be implemented in a future work:
- Priority to alocate slots by use count of underlying chunk. This heuristic
  increases probability of success reclaimation
- Varied chunk size. No need to allocate 32Mb chunk to split it in to 4K slots.
  It also increases probability of success reclaimation and improve perfomance of it