YQL-20436: Support Named Nodes on YqlSelect
This patch unlocks the last TPC-H query Q15, but has a couple of known
issues.

The first issue is that a named select with external column references
will work at a substitution place. To fix this I will introduce a new
`SqlSelect` option called `unknowns_disallowed`. I guess there is no
such issue in a pgSQL syntax, because there `WITH CTE` subqueries can
be referenced only from a `FROM` at a syntax level (CTE is not a column
reference).

The second issue is that `AutoYqlSelect` is broken, when `YqlSelect` is
used as a legacy `ISource`.
commit_hash:2c83192921d1f50899063b3b22461b1bf111b47e