YQL-20436: Fix asterisk subquery on SqlSelect
I explored the following problem:

```yql
PRAGMA YqlSelect = 'force';

/* BAD */ SELECT (SELECT * FROM (SELECT 1 AS a) AS x);

/* OK  */ SELECT (SELECT * FROM (SELECT 1 AS a) AS x);
          FROM (VALUES (1)) AS y (b);
```

There was a fatal error of not found qualified column name in a list of
unqualified column names, or vice versa. The condition for qualification
was dependent on a fact of SubLink type annotation. But initially
columns are qualified depending on existence of an external input.

Actually it seems that the existence of an external input is a right
condition for column qualification.
commit_hash:d03011a0f70ee94bf05418326cd4b6c655c5999f