YQL-20530: Fix a fatal table alias ending with '/'
I found a case, where the completion engine dies, when a table alias
ends with a `/` character.

The general problem is that the huge part of NameService stuff is built
on a fact that `/` is used as path delimeter. So when it is used at a
name, that will be unsafely contatenated somewhere with an other path,
strange things can happen.
commit_hash:69be32117450bfebf9609326f3a9ccef12985d61